# backend/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PulseFit Gym API",
    "version": "1.0.0",
    "description": "API Documentation for PulseFit Gym Management System"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/plans": {
      "get": {
        "tags": ["Plans"],
        "summary": "Get all active plans (pass ?all=true for all plans)",
        "security": [],
        "parameters": [
          { "name": "all", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "List of plans" }
        }
      },
      "post": {
        "tags": ["Plans"],
        "summary": "Create a new plan (Admin only)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "price": { "type": "number" },
                  "durationDays": { "type": "number" },
                  "category": { "type": "string", "enum": ["Drop-in", "Monthly", "Annual"] },
                  "features": { "type": "array", "items": { "type": "string" } },
                  "isActive": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Plan created" },
          "403": { "description": "Admin access required" }
        }
      }
    },
    "/plans/{id}": {
      "get": {
        "tags": ["Plans"],
        "summary": "Get a specific plan by ID",
        "security": [],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Plan data" },
          "404": { "description": "Plan not found" }
        }
      },
      "put": {
        "tags": ["Plans"],
        "summary": "Update a plan (Admin only)",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "price": { "type": "number" },
                  "durationDays": { "type": "number" },
                  "category": { "type": "string", "enum": ["Drop-in", "Monthly", "Annual"] },
                  "features": { "type": "array", "items": { "type": "string" } },
                  "isActive": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Plan updated" },
          "404": { "description": "Plan not found" },
          "403": { "description": "Admin access required" }
        }
      },
      "delete": {
        "tags": ["Plans"],
        "summary": "Delete a plan (Admin only)",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Plan deleted" },
          "404": { "description": "Plan not found" },
          "403": { "description": "Admin access required" }
        }
      }
    },
    "/subscriptions": {
      "post": {
        "tags": ["Subscriptions"],
        "summary": "Subscribe to a plan",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "planId": { "type": "string" }
                },
                "required": ["planId"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Subscription successful" },
          "400": { "description": "Already have an active plan" }
        }
      },
      "get": {
        "tags": ["Subscriptions"],
        "summary": "Get all subscriptions (Admin only)",
        "responses": {
          "200": { "description": "List of all subscriptions" },
          "403": { "description": "Admin access required" }
        }
      }
    },
...
    "/subscriptions/{id}/cancel": {
      "patch": {
        "tags": ["Subscriptions"],
        "summary": "Cancel a subscription",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Subscription cancelled" },
          "400": { "description": "Subscription is not active" },
          "404": { "description": "Subscription not found" }
        }
      }
    },
...
  }
}